put business logic to Post module
