+++ lib/nyon/web/controllers/post_controller.ex
@@ -1,13 +1,8 @@
 defmodule Nyon.Web.PostController do
   use Nyon.Web, :controller
 
-  @cache {:global, :posts}
-
   def index(conn, _params) do
-    posts = @cache
-      |> ConCache.ets
-      |> :ets.tab2list
-      |> Enum.map(fn {_key, post} -> post end)
+    posts = Nyon.Post.all
 
     conn
     |> put_status(200)
@@ -15,21 +10,10 @@ defmodule Nyon.Web.PostController do
   end
 
   def create(conn, %{"body" => body}) do
-    trimmed = String.trim(body)
-    post = %{"body" => trimmed}
-
-    if String.length(trimmed) > 0 do
-      ConCache.put(@cache, random_string(), post)
-    end
+    post = Nyon.Post.store(body)
 
     conn
     |> put_status(201)
     |> render("show.json", %{post: post})
-  end
-
-  defp random_string do
-    :crypto.strong_rand_bytes(64)
-    |> Base.url_encode64
-    |> binary_part(0, 64)
   end
 end
